src/requirements.txt
Flask
azure-identity
azure-storage-blob
python-dotenv
